+++ lib/presentation/pages/article_page/article_page.dart
@@ -2,13 +2,13 @@ import 'package:auto_route/auto_route.dart';
 import 'package:blogify_flutter_main/common/app_colors.dart';
 import 'package:blogify_flutter_main/common/app_dimensions.dart';
 import 'package:blogify_flutter_main/common/app_text_styles.dart';
-import 'package:blogify_flutter_main/data/mock_storage/global_mock_storage_provider.dart';
 import 'package:blogify_flutter_main/l10n/l10n.dart';
 import 'package:blogify_flutter_main/router/router.gr.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:provider/provider.dart';
 
+import '../../../data/providers/global_mock_storage_provider.dart';
 import '../../../utils/intl_formatter.dart';
 import '../../widgets/circled_button_outlined.dart';
 import '../../widgets/post_cover_photo.dart';
