+++ lib/presentation/pages/comments_page/comments_page.dart
@@ -2,9 +2,6 @@ import 'package:auto_route/auto_route.dart';
 import 'package:blogify_flutter_main/common/app_colors.dart';
 import 'package:blogify_flutter_main/common/app_dimensions.dart';
 import 'package:blogify_flutter_main/common/app_text_styles.dart';
-import 'package:blogify_flutter_main/data/mock_storage/global_mock_comment_provider.dart';
-import 'package:blogify_flutter_main/data/mock_storage/global_mock_storage_provider.dart';
-import 'package:blogify_flutter_main/data/mock_storage/global_mock_user_provider.dart';
 import 'package:blogify_flutter_main/domain/entities/comment_entity.dart';
 import 'package:blogify_flutter_main/domain/entities/post_entity.dart';
 import 'package:blogify_flutter_main/l10n/l10n.dart';
@@ -16,6 +13,9 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 
+import '../../../data/providers/global_mock_comment_provider.dart';
+import '../../../data/providers/global_mock_storage_provider.dart';
+import '../../../data/providers/global_mock_user_provider.dart';
 import '../../widgets/circled_button_outlined.dart';
 
 @RoutePage()
